feat!: Extracted handling of interactions into separate files
